+++ spec/lib/fields/polymorphic_spec.rb
@@ -1,4 +1,5 @@
 require "administrate/fields/polymorphic"
+require "support/constant_helpers"
 require "support/field_matchers"
 
 describe Administrate::Field::Polymorphic do
@@ -16,4 +17,24 @@ describe Administrate::Field::Polymorphic do
   end
 
   it { should_permit_param(:foo, for_attribute: :foo) }
+
+  describe "#display_associated_resource" do
+    it "displays through the dashboard based on the polymorphic class name" do
+      begin
+        Thing = Class.new
+        ThingDashboard = Class.new do
+          def display_resource(*)
+            :success
+          end
+        end
+
+        field = Administrate::Field::Polymorphic.new(:foo, Thing.new, :show)
+        display = field.display_associated_resource
+
+        expect(display).to eq :success
+      ensure
+        remove_constants :Thing, :ThingDashboard
+      end
+    end
+  end
 end